refactor: MVC 패턴 적용
App(controller)에서 입출력은 view로 분리, car는 model로 분리
